Feat: Add implementation of API POST:"/v2/roles"

+++ sqle/api/controller/v1/operation.go
@@ -34,7 +34,7 @@ func GetOperations(c echo.Context) error {
 
 	for i := range opCodes {
 		respData[i] = &OperationResV1{
-			Code: uint(opCodes[i]),
+			Code: opCodes[i],
 			Desc: model.GetOperationCodeDesc(opCodes[i]),
 		}
 	}
